+++ packages/amplify-cli/src/conditional-local-logging-init.ts
@@ -1,8 +1,8 @@
-import { CommandLineInput } from 'amplify-cli-core';
 import { JSONUtilities } from 'amplify-cli-core';
+import { CLIInput } from './domain/command-input';
 import { getAmplifyLogger, Redactor } from '@aws-amplify/amplify-cli-logger';
 
-export function logInput(input: CommandLineInput): void {
+export function logInput(input: CLIInput): void {
   getAmplifyLogger().logInfo({
     message: `amplify ${input.command ? input.command : ''} \
 ${input.plugin ? input.plugin : ''} \