+++ packages/amplify-cli/src/extensions/amplify-helpers/open-editor.ts
@@ -1,12 +1,11 @@
 import * as fs from 'fs-extra';
 import * as which from 'which';
-import open from 'open';
 import execa, { sync as execaSync } from 'execa';
 import * as inquirer from 'inquirer';
 import * as envEditor from 'env-editor';
 import { editorSelection } from './editor-selection';
 import { getEnvInfo } from './get-env-info';
-import { $TSContext } from 'amplify-cli-core';
+import { $TSContext, open } from 'amplify-cli-core';
 
 export async function openEditor(context: $TSContext, filePath: string, waitToContinue = true): Promise<void> {
   const continueQuestion: inquirer.InputQuestion = {