+++ packages/amplify-category-analytics/CHANGELOG.md
@@ -427,6 +427,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [4.4.2](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@4.4.1...amplify-category-analytics@4.4.2) (2023-02-16)
+
+**Note:** Version bump only for package amplify-category-analytics
+
+
+
+
+
 ## [4.4.1](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-analytics@4.4.0...amplify-category-analytics@4.4.1) (2023-02-06)
 
 **Note:** Version bump only for package amplify-category-analytics