+++ packages/amplify-category-hosting/CHANGELOG.md
@@ -251,6 +251,14 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
+## [3.4.10](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@3.4.9...amplify-category-hosting@3.4.10) (2023-02-16)
+
+**Note:** Version bump only for package amplify-category-hosting
+
+
+
+
+
 ## [3.4.9](https://github.com/aws-amplify/amplify-cli/compare/amplify-category-hosting@3.4.8...amplify-category-hosting@3.4.9) (2023-02-06)
 
 **Note:** Version bump only for package amplify-category-hosting